requirements.txt
astroid==2.4.2
colorama==0.4.4
greenlet==2.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.5
parso==0.7.1
pygame==2.4
pynvim==0.4.2
six==1.15.0
toml==0.10.2
wrapt==1.12.1
neat-python==0.92
python-dotenv==0.15.0